requirements.txt
boto3>=1.28.0
langchain>=0.0.267
langchain-community>=0.0.10
opensearch-py>=2.2.0
requests-aws4auth>=1.2.3
pymupdf>=1.22.5
pandas>=2.0.0
pillow>=10.0.0
numpy>=1.24.0
uuid>=1.30
